Release 0.3.0 (#115)
* feat(api-server): create `llama-core`


* chore(core): remove `ModelInfo`


* chore(core): update `chat` and `models`


* feat(api-server): integrate `llama-core`


* refactor(core): update `Graph` and `init_core_context`


* refactor(core): update `models` API


* refactor(llama-api-server): update code


* refactor(endpoints): update embeddings to be compatible with OpenAI API 2.0


* feat(llama-api-server): enable `/v1/embeddings`


* chore(endpoints): update `EmbeddingRequest`


* feat(endpoints): define `EmbeddingsResponse` in `embeddings`


* chore(llama-api-server): update `embeddings_handler`


* chore(llama-api-server): update code


* feat(core): add `embeddings`


* refactor(llama-api-server): integrate `embeddings` API of `llama-core`


* refactor!(endpoints): update the type of `EmbeddingObject::index` field


* refactor!(core): improve `embeddings` API


* chore(llama-api-server): adapt to the `embeddings` API of `llama-core`


* refactor!(endpoints): update `Usage`


* refactor!(endpoints): update `EmbeddingsResponse`


* chore(llama-api-server): update code


* refactor!(core): update `embeddings` API


* chore(llama-api-server): update code


* chore(llama-api-server): remove debug code


* chore(core): remove debug code


* version(endpoints): bump to `0.4.0`


* chore(api-server): update dependency


* chore(chat): update dependency


* feat!(core): update stream terminated string


* chore(chat): improve the utf8 decoding


* chore(core): update the checking for the length of `cached_encodings`


* chore(llama-api-server): update `DEFAULT_SOCKET_ADDRESS`


* doc: update `README` of `chat-prompts` and `models.md` (#114)

* Update models.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* version: bump to `0.3.0`


---------

+++ api-server/endpoints/src/common.rs
@@ -7,14 +7,14 @@ pub enum LlamaCppLogitBiasType {
     tokens,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Default, Deserialize, Serialize)]
 pub struct Usage {
     /// Number of tokens in the prompt.
-    pub prompt_tokens: u32,
+    pub prompt_tokens: u64,
     /// Number of tokens in the generated completion.
-    pub completion_tokens: u32,
+    pub completion_tokens: u64,
     /// Total number of tokens used in the request (prompt + completion).
-    pub total_tokens: u32,
+    pub total_tokens: u64,
 }
 
 /// The reason the model stopped generating tokens.
